+++ wxc_sdk/as_api.py
@@ -1266,7 +1266,7 @@ class AsDevicesApi(AsApiChild, base='devices'):
         if model is not None:
             body['model'] = model
         if password is not None:
-            body.password = password
+            body['password'] = password
         url = self.ep()
         data = await super().post(url=url, json=body, params=params)
         return Device.model_validate(data)